fix: satisfy browser annotation ci checks
