+++ test/unit/features/browser/browser_annotation_test.dart
@@ -2,6 +2,39 @@ import 'package:alera/src/features/browser/domain/browser_annotation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
+  test('serializes anchors and preserves comment copies', () {
+    const anchor = BrowserAnnotationAnchor(
+      x: -0.2,
+      y: 0.4,
+      width: 1.4,
+      height: 0.5,
+      role: 'button',
+      name: 'Save',
+      tag: 'button',
+    );
+    expect(anchor.geometry, '0.0%,40.0%,100.0%,50.0%');
+    expect(anchor.toJson(), <String, Object?>{
+      'x': -0.2,
+      'y': 0.4,
+      'width': 1.4,
+      'height': 0.5,
+      'role': 'button',
+      'name': 'Save',
+      'tag': 'button',
+    });
+    expect(const BrowserAnnotationElement(anchor: anchor).anchor, anchor);
+
+    const comment = BrowserAnnotationComment(
+      id: 'comment',
+      kind: BrowserAnnotationKind.element,
+      anchor: anchor,
+      text: 'Old text',
+    );
+    expect(comment.copyWith(text: 'New text').text, 'New text');
+    expect(comment.copyWith().text, 'Old text');
+    expect(comment.toJson(index: 3)['index'], 3);
+  });
+
   group('BrowserAnnotationCapture', () {
     test('builds structured context for element and region comments', () {
       final capture = BrowserAnnotationCapture(
@@ -73,6 +106,7 @@ void main() {
       );
 
       final json = capture.toJson();
+      expect(capture.displayName, 'Browser Annotation (1 Comment)');
       expect(json['url'], 'https://example.com');
       expect(json['capturedAt'], '2026-08-12T00:00:00.000Z');
       expect((json['comments'] as List).single, <String, Object?>{
@@ -83,5 +117,54 @@ void main() {
         'text': 'Review this area.',
       });
     });
+
+    test(
+      'uses fallback element labels and supports empty and copied captures',
+      () {
+        const anchors = <BrowserAnnotationAnchor>[
+          BrowserAnnotationAnchor(x: 0, y: 0, width: 0, height: 0),
+          BrowserAnnotationAnchor(
+            x: 0,
+            y: 0,
+            width: 0,
+            height: 0,
+            name: 'Settings',
+          ),
+          BrowserAnnotationAnchor(
+            x: 0,
+            y: 0,
+            width: 0,
+            height: 0,
+            tag: 'section',
+          ),
+        ];
+        final capture = BrowserAnnotationCapture(
+          imagePath: '/tmp/annotation.png',
+          url: Uri.parse('https://example.com'),
+          title: '',
+          viewportWidth: 1,
+          viewportHeight: 1,
+          capturedAt: DateTime.utc(2026, 8, 12),
+          comments: <BrowserAnnotationComment>[
+            for (var index = 0; index < anchors.length; index++)
+              BrowserAnnotationComment(
+                id: '$index',
+                kind: BrowserAnnotationKind.element,
+                anchor: anchors[index],
+                text: 'Comment $index',
+              ),
+          ],
+        );
+        expect(capture.displayName, 'Browser Annotation (3 Comments)');
+        expect(capture.contextText, contains('Element: unknown'));
+        expect(capture.contextText, contains('Element: "Settings"'));
+        expect(capture.contextText, contains('Element: section'));
+
+        final empty = capture.copyWith(comments: const []);
+        expect(empty.displayName, 'Browser Annotation (0 Comments)');
+        expect(empty.contextText, contains('Page: example.com'));
+        expect(capture.copyWith().comments, capture.comments);
+      },
+    );
   });
 }
